+++ common/src/main/java/com/nemonotfound/nemos/copper/item/ModCreativeModeTabs.java
@@ -24,7 +24,12 @@ public class ModCreativeModeTabs {
                 output.accept(ModItems.COPPER_CHEST_MINECART.get());
                 output.accept(ModItems.COPPER_FURNACE_MINECART.get());
                 output.accept(ModItems.COPPER_TNT_MINECART.get());
+                output.accept(ModItems.COPPER_SHOVEL.get());
+                output.accept(ModItems.COPPER_PICKAXE.get());
+                output.accept(ModItems.COPPER_AXE.get());
+                output.accept(ModItems.COPPER_HOE.get());
                 output.accept(ModItems.COPPER_SHEARS.get());
+                output.accept(ModItems.COPPER_SWORD.get());
             })
             .build());