+++ Changelog.md
@@ -2,3 +2,8 @@
 
 ## Additions
 - Added Copper Shears
+- Added Copper Shovel
+- Added Copper Pickaxe
+- Added Copper Axe
+- Added Copper Hoe
+- Added Copper Sword